macro simplification, show types in insight, source maps for dev and prod
simplify macro instrumentation to the bare minimum
apply, select, ident, block, for/yield, match, type apply
removing val/var right hand side

+++ codemirror/src/main/scala/org.denigma.codemirror/Editor.scala
@@ -96,9 +96,9 @@ class Doc protected () extends js.Object {
 }
 
 trait Coords extends js.Object {
-  val left: Int = js.native
-  val top: Int = js.native
-  val bottom: Int = js.native
+  val left: Double = js.native
+  val top: Double = js.native
+  val bottom: Double = js.native
 }
 
 trait Range extends js.Object {
